Prepare for releasing
Fixes some static analysis issues
Sets version number to 0.2.0
Adds provisioning profiles for distribution and development

+++ Werewolf Moderator/Werewolf Moderator/PlayControllers/Ballot/BallotSecondRoundViewController.m
@@ -39,7 +39,7 @@
 
 - (IBAction)voteStep {
     self.currentVoteCount = self.voteStepper.value;
-    self.voteCountLabel.text = [NSString stringWithFormat:@"%ld", self.currentVoteCount];
+    self.voteCountLabel.text = [NSString stringWithFormat:@"%ld", (long)self.currentVoteCount];
 }
 
 - (IBAction)submitVotes {
